CSS: add paragraph numbering at fixed position
Paragraph counter is reset on every header, and its position is fixed.
Preferably position of paragraph number should be dynamically
determined, so that paragraph number would be located in the ~same
distance to paragraph, and its number would not overlap with the
paragraph.

Also add bottom margin to the tables.